Module tblgen::raw

This module contains raw bindings for TableGen. Note that these bindings are unstable and can change at any time.

Modules

TableGenDiagKind

TableGenRecTyKind

Structs

TableGen

TableGenDagPair

TableGenRecord

TableGenRecordArray

TableGenRecordKeeper

TableGenRecordMap

TableGenRecordMapIterator

TableGenRecordVal

TableGenRecordValArray

TableGenRecordVector

TableGenSourceLocation

TableGenStringRef

TableGenTypedInit

__fsid_t

max_align_t

Constants

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

PTRDIFF_MAX

PTRDIFF_MIN

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WINT_MAX

WINT_MIN

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_FEATURES_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_SYS_CDEFS_H

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_IEC_60559_BFP__

__STDC_IEC_60559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__glibc_c99_flexarr_available

Functions

tableGenAddIncludePath⚠

tableGenAddSource⚠

tableGenAddSourceFile⚠

tableGenBitArrayFree⚠

tableGenBitInitGetValue⚠

tableGenBitsInitGetBitInit⚠

tableGenBitsInitGetNumBits⚠

tableGenBitsInitGetValue⚠

tableGenDagRecordArgName⚠

tableGenDagRecordGet⚠

tableGenDagRecordNumArgs⚠

tableGenDagRecordOperator⚠

tableGenDefInitGetValue⚠

tableGenFree⚠

tableGenGet⚠

tableGenInitDump⚠

tableGenInitPrint⚠

tableGenInitRecType⚠

tableGenIntInitGetValue⚠

tableGenListRecordGet⚠

tableGenListRecordGetType⚠

tableGenListRecordNumElements⚠

tableGenParse⚠

NOTE: TableGen currently relies on global state within a given parser invocation, so this function is not thread-safe.

tableGenPrintError⚠

tableGenRecordDump⚠

tableGenRecordGetFieldType⚠

tableGenRecordGetFirstValue⚠

tableGenRecordGetLoc⚠

tableGenRecordGetName⚠

tableGenRecordGetRecords⚠

tableGenRecordGetValue⚠

tableGenRecordIsAnonymous⚠

tableGenRecordIsSubclassOf⚠

tableGenRecordKeeperFree⚠

tableGenRecordKeeperGetAllDerivedDefinitions⚠

tableGenRecordKeeperGetClass⚠

tableGenRecordKeeperGetClasses⚠

tableGenRecordKeeperGetDef⚠

tableGenRecordKeeperGetDefs⚠

tableGenRecordKeeperGetFirstClass⚠

tableGenRecordKeeperGetFirstDef⚠

tableGenRecordKeeperGetNextClass⚠

tableGenRecordKeeperGetNextDef⚠

tableGenRecordKeeperItemGetName⚠

tableGenRecordKeeperItemGetRecord⚠

tableGenRecordKeeperIteratorClone⚠

tableGenRecordKeeperIteratorFree⚠

tableGenRecordPrint⚠

tableGenRecordValDump⚠

tableGenRecordValGetLoc⚠

tableGenRecordValGetName⚠

tableGenRecordValGetNameInit⚠

tableGenRecordValGetType⚠

tableGenRecordValGetValAsBit⚠

tableGenRecordValGetValAsBits⚠

tableGenRecordValGetValAsDefRecord⚠

tableGenRecordValGetValAsInt⚠

tableGenRecordValGetValAsNewString⚠

tableGenRecordValGetValAsRecord⚠

tableGenRecordValGetValue⚠

tableGenRecordValNext⚠

tableGenRecordValPrint⚠

tableGenRecordValTest⚠

tableGenRecordVectorFree⚠

tableGenRecordVectorGet⚠

tableGenSourceLocationClone⚠

tableGenSourceLocationFree⚠

tableGenSourceLocationNull⚠

tableGenStringArrayFree⚠

tableGenStringFree⚠

tableGenStringInitGetValue⚠

tableGenStringInitGetValueNewString⚠

Type Aliases

TableGenBool

TableGenDagPairRef

TableGenParserRef

TableGenRecordArrayRef

TableGenRecordKeeperIteratorRef

TableGenRecordKeeperRef

TableGenRecordMapRef

TableGenRecordRef

TableGenRecordValArrayRef

TableGenRecordValRef

TableGenRecordVectorRef

TableGenSourceLocationRef

TableGenStringCallback

TableGenTypedInitRef

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

wchar_t