Struct tblgen::record::Record

pub struct Record<'a> { /* private fields */ }

An immutable reference to a TableGen record.

This reference cannot outlive the RecordKeeper from which it is borrowed.

Implementations

impl<'a> Record<'a>

pub unsafe fn from_raw(ptr: TableGenRecordRef) -> Record<'a>

Creates a record from a raw object.

Safety

The raw object must be valid.

pub fn name(self) -> Result<&'a str, Error>

Returns the name of the record.

Errors

Returns an error if the name is not a valid UTF-8 string.

pub fn bit_value<'n>(self, name: &'n str) -> Result<bool, Error>

Returns the boolean value of the field with the given name if this field is of type BitInit.

pub fn bits_value<'n>(self, name: &'n str) -> Result<Vec<bool>, Error>

Returns the field with the given name converted to a Vec<bool> if this field is of type BitsInit.

pub fn int_value<'n>(self, name: &'n str) -> Result<i64, Error>

Returns the integer value of the field with the given name if this field is of type IntInit.

pub fn code_value<'n>(self, name: &'n str) -> Result<String, Error>

Returns the field with the given name converted to a String if this field is of type StringInit.

Note that this copies the string into a new string.

pub fn code_str_value<'n>(self, name: &'n str) -> Result<&'a str, Error>

Returns the field with the given name converted to a &str if this field is of type StringInit.

pub fn string_value<'n>(self, name: &'n str) -> Result<String, Error>

Returns the field with the given name converted to a String if this field is of type StringInit.

Note that this copies the string into a new string.

pub fn str_value<'n>(self, name: &'n str) -> Result<&'a str, Error>

Returns the field with the given name converted to a &str if this field is of type StringInit.

pub fn def_value<'n>(self, name: &'n str) -> Result<Record<'a>, Error>

Returns the field with the given name converted to a Record if this field is of type DefInit.

pub fn list_value<'n>(self, name: &'n str) -> Result<ListInit<'a>, Error>

Returns the field with the given name converted to a ListInit if this field is of type ListInit.

pub fn dag_value<'n>(self, name: &'n str) -> Result<DagInit<'a>, Error>

Returns the field with the given name converted to a DagInit if this field is of type DagInit.

pub fn value<'n>(self, name: &'n str) -> Result<RecordValue<'a>, Error>

Returns a RecordValue for the field with the given name.

pub fn anonymous(self) -> bool

Returns true if the record is anonymous.

pub fn subclass_of(self, class: &str) -> bool

Returns true if the record is a subclass of the class with the given name.

pub fn values(self) -> RecordValueIter<'a>

Returns an iterator over the fields of the record.

The iterator yields RecordValue structs